refactored SearchResult<T> into common/data-model/solr-data-model
