+++ src/main/java/eu/nimble/indexing/service/CatalogueService.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 import org.springframework.data.domain.Pageable;
 
-import eu.nimble.indexing.model.IndexField;
-import eu.nimble.indexing.model.SearchResult;
+import eu.nimble.service.model.solr.IndexField;
+import eu.nimble.service.model.solr.SearchResult;
 import eu.nimble.service.model.solr.item.ItemType;
 
 public interface CatalogueService {
