+++ src/main/java/eu/nimble/indexing/service/impl/ClassServiceImpl.java
@@ -15,9 +15,9 @@ import org.springframework.data.solr.core.query.SimpleQuery;
 import org.springframework.data.solr.core.query.result.ScoredPage;
 import org.springframework.stereotype.Service;
 
-import eu.nimble.indexing.model.SearchResult;
 import eu.nimble.indexing.repository.ClassRepository;
 import eu.nimble.indexing.service.ClassService;
+import eu.nimble.service.model.solr.SearchResult;
 import eu.nimble.service.model.solr.owl.ClassType;
 import eu.nimble.service.model.solr.owl.IClassType;
 import eu.nimble.service.model.solr.owl.IPropertyType;
